# gateway/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Payment Gateway API
  version: 1.0.0
  summary: Public contract for gateway-owned Payment operations.
  description: |
    The Payment Gateway API creates and manages gateway-owned Payments between an
    e-commerce order service and the Mock Bank. Payment IDs use the `pay_<uuid>`
    form. Amounts are expressed in cents and Currency is always `USD`.

    The runnable request collection in `demo/payment-gateway.http` remains the
    companion artifact for trying this contract against the local demo stack.
servers:
  - url: http://localhost:8080
    description: Local root demo stack
tags:
  - name: Payments
security:
  - BearerAuth: []
paths:
  /v1/payments:
    post:
      tags:
        - Payments
      operationId: authorizePayment
      summary: Create and authorize a Payment
      description: |
        Requires the `payments:write` scope. Creates a Payment and asks the
        Mock Bank to authorize the full Amount.
        The response Payment Status can be `authorized`, `expired`,
        `declined`, or `pending`. `pending` means the Mock Bank accepted the
        authorization request but did not produce a definitive approval or
        decline outcome. Mock Bank dependency failures before a usable bank
        response are returned as public `bank_unavailable` or `bank_timeout`
        errors. If the synchronous command exhausts its overall deadline, it
        returns `payment_timeout`; this does not mean that the Payment failed.
        Retry either timeout with the same Idempotency Key.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizePaymentRequest"
            examples:
              authorize:
                value:
                  order_id: order-1001
                  customer_id: customer-501
                  amount: 1299
                  card:
                    number: "4111111111111111"
                    cvv: "123"
                    expiry_month: 12
                    expiry_year: 2030
      responses:
        "201":
          description: Payment was created.
          headers:
            Location:
              description: Path of the created Payment resource.
              schema:
                type: string
                example: /v1/payments/pay_550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "200":
          description: Idempotency Replay for a completed authorize request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/InvalidJsonBody"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BankDependencyFailure"
        "504":
          $ref: "#/components/responses/PaymentCommandTimeout"
    get:
      tags:
        - Payments
      operationId: searchPayments
      summary: Search Payments
      description: Search Payments by external Order ID or Customer ID, optionally filtered by Payment Status. Requires the `payments:read` scope.
      parameters:
        - name: order_id
          in: query
          description: External Order ID owned by the order service. Required when `customer_id` is omitted.
          schema:
            type: string
            example: order-1001
        - name: customer_id
          in: query
          description: External Customer ID. Required when `order_id` is omitted.
          schema:
            type: string
            example: customer-501
        - name: status
          in: query
          description: Optional public Payment Status filter.
          schema:
            $ref: "#/components/schemas/PaymentStatus"
      responses:
        "200":
          description: Matching Payments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/ValidationError"
        "504":
          $ref: "#/components/responses/PaymentReadTimeout"
  /v1/payments/{payment_id}:
    get:
      tags:
        - Payments
      operationId: fetchPayment
      summary: Fetch a Payment
      description: Requires the `payments:read` scope.
      parameters:
        - $ref: "#/components/parameters/PaymentID"
      responses:
        "200":
          description: Payment found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "404":
          $ref: "#/components/responses/PaymentNotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "504":
          $ref: "#/components/responses/PaymentReadTimeout"
  /v1/payments/{payment_id}/authorization-retries:
    post:
      tags:
        - Payments
      operationId: retryAuthorization
      summary: Retry a Pending Payment authorization
      description: |
        Requires the `payments:write` scope. Retries authorization for a
        Payment whose current Payment Status is `pending`. The Authorization
        Retry must use the same card number and expiry as the original Payment
        authorization request.
      parameters:
        - $ref: "#/components/parameters/PaymentID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryAuthorizationRequest"
      responses:
        "200":
          description: Authorization Retry completed or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/InvalidJsonBody"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "404":
          $ref: "#/components/responses/PaymentNotFound"
        "409":
          $ref: "#/components/responses/PaymentCommandConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BankDependencyFailure"
        "504":
          $ref: "#/components/responses/PaymentCommandTimeout"
  /v1/payments/{payment_id}/capture:
    post:
      tags:
        - Payments
      operationId: capturePayment
      summary: Capture a Payment
      description: |
        Requires the `payments:write` scope. Captures the full Amount for an
        `authorized` Payment. This command has no input fields; callers must
        send an empty request body. A non-empty request body is rejected as
        `invalid_json_body`.
      parameters:
        - $ref: "#/components/parameters/PaymentID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/EmptyCommandBody"
      responses:
        "200":
          description: Payment captured or Idempotency Replay returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/InvalidJsonBody"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "404":
          $ref: "#/components/responses/PaymentNotFound"
        "409":
          $ref: "#/components/responses/PaymentCommandConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BankDependencyFailure"
        "504":
          $ref: "#/components/responses/PaymentCommandTimeout"
  /v1/payments/{payment_id}/void:
    post:
      tags:
        - Payments
      operationId: voidPayment
      summary: Void a Payment
      description: |
        Requires the `payments:write` scope. Cancels the full authorization for
        an `authorized` Payment before it is captured. This command has no
        input fields; callers must send an empty request body. A non-empty
        request body is rejected as `invalid_json_body`.
      parameters:
        - $ref: "#/components/parameters/PaymentID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/EmptyCommandBody"
      responses:
        "200":
          description: Payment voided or Idempotency Replay returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/InvalidJsonBody"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "404":
          $ref: "#/components/responses/PaymentNotFound"
        "409":
          $ref: "#/components/responses/PaymentCommandConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BankDependencyFailure"
        "504":
          $ref: "#/components/responses/PaymentCommandTimeout"
  /v1/payments/{payment_id}/refund:
    post:
      tags:
        - Payments
      operationId: refundPayment
      summary: Refund a Payment
      description: |
        Requires the `payments:write` scope. Returns the full Amount for a
        `captured` Payment. This command has no input fields; callers must send
        an empty request body. A non-empty request body is rejected as
        `invalid_json_body`.
      parameters:
        - $ref: "#/components/parameters/PaymentID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/EmptyCommandBody"
      responses:
        "200":
          description: Payment refunded or Idempotency Replay returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "400":
          $ref: "#/components/responses/InvalidJsonBody"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "404":
          $ref: "#/components/responses/PaymentNotFound"
        "409":
          $ref: "#/components/responses/PaymentCommandConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BankDependencyFailure"
        "504":
          $ref: "#/components/responses/PaymentCommandTimeout"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque Service Credential
      description: Opaque Order Service credential. Payment reads require `payments:read`; Payment commands require `payments:write`.
  parameters:
    PaymentID:
      name: payment_id
      in: path
      required: true
      description: Gateway-owned Payment ID.
      schema:
        type: string
        pattern: "^pay_[0-9a-fA-F-]{36}$"
        example: pay_550e8400-e29b-41d4-a716-446655440000
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-provided opaque operation identity for mutating commands. It must
        be non-empty. Reusing the same key for the same operation and same
        request values returns an Idempotency Replay. Reusing it with different
        request values returns `409 idempotency_key_conflict`.
      schema:
        type: string
        minLength: 1
        example: authorize-order-1001
  requestBodies:
    EmptyCommandBody:
      required: false
      description: The request body must be empty. Do not send `{}` or JSON `null`.
      content: {}
  responses:
    RateLimited:
      description: Payment API quota is exhausted. Retry after the stated whole-second delay.
      headers:
        Retry-After:
          required: true
          description: Whole seconds until the next quota token is available.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: rate_limited
              message: rate limit exceeded
    Unauthorized:
      description: Missing, malformed, or invalid Bearer Service Credential.
      headers:
        WWW-Authenticate:
          required: true
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: Unauthorized
    Forbidden:
      description: Authenticated Service Principal lacks the required Payment Scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: forbidden
              message: Forbidden
    InvalidJsonBody:
      description: Request JSON is malformed, invalid for the endpoint, too large, has unknown fields, has multiple values, or a no-input command received a non-empty body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidJsonBody:
              value:
                error:
                  code: invalid_json_body
                  message: invalid JSON body
    RequestBodyTooLarge:
      description: Request body exceeds the configured gateway limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            requestBodyTooLarge:
              value:
                error:
                  code: request_body_too_large
                  message: request body is too large
    UnsupportedMediaType:
      description: |
        JSON request endpoints require `Content-Type: application/json`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unsupportedMediaType:
              value:
                error:
                  code: unsupported_media_type
                  message: content type must be application/json
    ValidationError:
      description: |
        Validation failed, including missing Idempotency Key, invalid Payment ID
        shape, invalid card details, invalid amount, unsupported search filters,
        or missing search identity filters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validationError:
              value:
                error:
                  code: validation_error
                  message: payment request is invalid
    PaymentNotFound:
      description: No Payment exists for the requested Payment ID.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            paymentNotFound:
              value:
                error:
                  code: payment_not_found
                  message: payment was not found
    IdempotencyConflict:
      description: The Idempotency Key is already in progress or was already used with different request values.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            keyConflict:
              value:
                error:
                  code: idempotency_key_conflict
                  message: idempotency key was already used with a different request
            keyInProgress:
              value:
                error:
                  code: idempotency_key_in_progress
                  message: idempotency key is already in progress
    PaymentCommandConflict:
      description: Payment Status conflict, authorization expiration, or Idempotency Key conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            statusConflict:
              value:
                error:
                  code: payment_status_conflict
                  message: payment status does not allow this operation
            keyConflict:
              value:
                error:
                  code: idempotency_key_conflict
                  message: idempotency key was already used with a different request
            keyInProgress:
              value:
                error:
                  code: idempotency_key_in_progress
                  message: idempotency key is already in progress
    BankDependencyFailure:
      description: Mock Bank dependency failure or bank state conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            bankUnavailable:
              value:
                error:
                  code: bank_unavailable
                  message: bank is unavailable
            bankStateConflict:
              value:
                error:
                  code: bank_state_conflict
                  message: bank state conflicts with local payment state
    PaymentCommandTimeout:
      description: "A timeout has one of two distinct outcomes: `bank_timeout` means the Mock Bank child budget expired while command time remained, and `payment_timeout` means the overall synchronous Payment command budget expired. Retry either with the same Idempotency Key; neither means the Payment failed."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            bankTimeout:
              value:
                error:
                  code: bank_timeout
                  message: bank request timed out
            paymentTimeout:
              value:
                error:
                  code: payment_timeout
                  message: payment command timed out; retry with the same idempotency key
    PaymentReadTimeout:
      description: The Payment read exceeded its request deadline.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            requestTimeout:
              value:
                error:
                  code: request_timeout
                  message: payment read timed out
  schemas:
    AuthorizePaymentRequest:
      type: object
      additionalProperties: false
      required:
        - order_id
        - customer_id
        - amount
        - card
      properties:
        order_id:
          type: string
          minLength: 1
          description: External Order ID owned by the order service.
          example: order-1001
        customer_id:
          type: string
          minLength: 1
          description: External Customer ID.
          example: customer-501
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Payment Amount in cents.
          example: 1299
        card:
          $ref: "#/components/schemas/CardDetails"
    RetryAuthorizationRequest:
      type: object
      additionalProperties: false
      required:
        - card
      properties:
        card:
          $ref: "#/components/schemas/CardDetails"
    CardDetails:
      type: object
      additionalProperties: false
      required:
        - number
        - cvv
        - expiry_month
        - expiry_year
      properties:
        number:
          type: string
          pattern: "^[0-9]{12,19}$"
          description: Card number sent to the Mock Bank. The gateway does not return it in public responses.
          example: "4111111111111111"
        cvv:
          type: string
          pattern: "^[0-9]{3,4}$"
          description: Card CVV sent to the Mock Bank. The gateway does not store or return it.
          example: "123"
        expiry_month:
          type: integer
          minimum: 1
          maximum: 12
          example: 12
        expiry_year:
          type: integer
          minimum: 1
          example: 2030
    PaymentEnvelope:
      type: object
      additionalProperties: false
      required:
        - payment
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
    PaymentsEnvelope:
      type: object
      additionalProperties: false
      required:
        - payments
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    Payment:
      type: object
      additionalProperties: false
      required:
        - id
        - order_id
        - customer_id
        - amount
        - currency
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Gateway-owned Payment ID.
          example: pay_550e8400-e29b-41d4-a716-446655440000
        order_id:
          type: string
          description: External Order ID owned by the order service.
          example: order-1001
        customer_id:
          type: string
          description: External Customer ID.
          example: customer-501
        amount:
          type: integer
          format: int64
          description: Payment Amount in cents.
          example: 1299
        currency:
          type: string
          const: USD
          description: Payment Currency. The gateway only uses US dollars.
          example: USD
        status:
          $ref: "#/components/schemas/PaymentStatus"
        decline_reason:
          type: string
          description: Gateway-owned explanation for a Declined Payment.
          example: insufficient_funds
        authorization_expires_at:
          type: string
          format: date-time
          description: Authorization Expiration Time for authorized, captured, voided, or refunded Payments when known.
          example: "2026-06-18T13:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2026-06-18T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-18T12:00:00Z"
    PaymentStatus:
      type: string
      description: |
        Public Payment Status:
        - `pending`: authorization outcome is not yet known because the Mock Bank has not produced a definitive approval or decline.
        - `authorized`: the full Amount has been reserved by the Mock Bank and can still be captured or voided.
        - `expired`: authorization was approved by the Mock Bank but can no longer be captured or voided because the authorization hold expired.
        - `declined`: authorization was refused by the Mock Bank.
        - `captured`: the full authorized Amount has been collected.
        - `voided`: authorization was cancelled before capture.
        - `refunded`: the full captured Amount has been returned.
      enum:
        - pending
        - authorized
        - expired
        - declined
        - captured
        - voided
        - refunded
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_json_body
                - unsupported_media_type
                - validation_error
                - idempotency_key_conflict
                - idempotency_key_in_progress
                - payment_status_conflict
                - payment_not_found
                - bank_state_conflict
                - unauthorized
                - forbidden
                - bank_unavailable
                - bank_timeout
                - payment_timeout
                - request_timeout
                - request_body_too_large
                - internal_server_error
                - service_unavailable
              example: payment_not_found
            message:
              type: string
              example: payment was not found